Extracted MemberService interface.
